move generate_markdown.py to dedicated docs/scripts folder
