+++ docs/source/conf.py
@@ -1,8 +1,8 @@
 import datetime
 import os
 import sys
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..')))
-from reeds import generate_markdown
+sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "scripts"))
+import generate_markdown
 
 # Configuration file for the Sphinx documentation builder.
 #
